+++ derive_builder/src/options/darling_opts.rs
@@ -294,6 +294,16 @@ impl Options {
             .fields
     }
 
+    /// A builder requires `Clone` to be derived if any of its fields use the mutable or immutable pattern.
+    ///
+    /// This can technically be expressed two ways:
+    ///
+    /// 1. The caller chooses the `owned` pattern at the struct level
+    /// 1. The caller chooses the `owned` pattern on every field
+    pub fn requires_clone(&self) -> bool {
+        self.fields().any(|f| f.pattern().requires_clone())
+    }
+
     /// Get an iterator over the input struct's fields which pulls fallback
     /// values from struct-level settings.
     pub fn fields<'a>(&'a self) -> FieldIter<'a> {
@@ -323,6 +333,7 @@ impl Options {
             visibility: self.builder_vis(),
             fields: Vec::with_capacity(self.field_count()),
             functions: Vec::with_capacity(self.field_count()),
+            must_derive_clone: self.requires_clone(),
             doc_comment: None,
             deprecation_notes: Default::default(),
             bindings: self.bindings(),